Be prepared for names, passwords containing space or quote
Some roles could fail if the configuration values contained
space or quote characters or were simply empty strings.

For example in client_join role if client_password configuration
value was an empty string then the role failed to have the client
joined to the server but did not notice any problem and reported
success.